+++ packages/@storefront/flux-capacitor/src/core/requests/utils.ts
@@ -182,6 +182,9 @@ namespace RequestHelpers {
 
   export const chain = <T>(...fns: Array<(...obj: any[]) => T>): T =>
     fns.reduce((final, fn) => fn(final) || final, <T>{});
+
+  export const attachSessionId = <T>(state: Store.State): ((r: T) => T & { sessionId: Store.SessionId }) =>
+    (r) => Object.assign({ sessionId: state.session.sessionId }, r);
 }
 
 export default RequestHelpers;