+++ packages/@storefront/flux-capacitor/test/unit/core/requests/index.ts
@@ -77,9 +77,8 @@ suite('RequestBuilder', ({ expect, stub, spy }) => {
  describe('composeRequest()', () => {
    it('should combine build, and override, and call setPastState', () => {
      const req = { a: 'b' };
-     const overrideFn = () => null;
-     const pastStateKey = 'search';
      const normalizedReq = { c: 'd' };
+     const attachSessionId = () => ({ g: 'h' });
      const oFn = (s) => s;
      const o = () => ({ e: 'f' });
      const setPastState = () => null;
@@ -91,10 +90,11 @@ suite('RequestBuilder', ({ expect, stub, spy }) => {
      requestBuilder.override = stub().withArgs(oFn, requestBuilder).returns(o);
      requestBuilder.setPastState = stub().withArgs(requestBuilder).returns(setPastState);
      stub(utils, 'normalizeToFunction').withArgs(req).returns(normalizedReq);
+     stub(RequestHelpers, 'attachSessionId').withArgs(state).returns(attachSessionId);
 
      requestBuilder.composeRequest(state, overrideRequest);
 
-     expect(requestChain).to.be.calledWithExactly(normalizedReq, o, setPastState);
+     expect(requestChain).to.be.calledWithExactly(normalizedReq, attachSessionId, o, setPastState);
    });
  });
 });